requirements.txt
opencv-python
pillow==10.2.0
tkvideoplayer==2.8
pywin32==306
requests~=2.31.0
fastapi~=0.109.2
uvicorn~=0.27.1
pydantic~=2.6.1
numpy~=1.26.3